MINIFICPP-1602 upgrade Catch2 to v2.13.6

+++ libminifi/test/unit/MemoryUsageTest.cpp
@@ -18,6 +18,7 @@
 
 #include <cstring>
 
+#include "utils/gsl.h"
 #include "utils/OsUtils.h"
 #include "../TestBase.h"
 
@@ -35,8 +36,8 @@ TEST_CASE("Test Physical memory usage", "[testphysicalmemoryusage]") {
     std::cout << "Physical Memory used by the system: " << ram_usage_by_system << v.data();
     std::cout << "Total Physical Memory in the system: " << ram_total << v.data();
   }
-  REQUIRE(ram_usage_by_process >= v.size());
-  REQUIRE(v.size()*2 >= ram_usage_by_process);
+  REQUIRE(ram_usage_by_process >= gsl::narrow<int64_t>(v.size()));
+  REQUIRE(gsl::narrow<int64_t>(v.size()*2) >= ram_usage_by_process);
   REQUIRE(ram_usage_by_system >= ram_usage_by_process);
   REQUIRE(ram_total >= ram_usage_by_system);
 }
@@ -47,7 +48,7 @@ size_t GetTotalMemoryLegacy() {
 }
 
 TEST_CASE("Test new and legacy total system memory query equivalency") {
-  REQUIRE(GetTotalMemoryLegacy() == utils::OsUtils::getSystemTotalPhysicalMemory());
+  REQUIRE(gsl::narrow<int64_t>(GetTotalMemoryLegacy()) == utils::OsUtils::getSystemTotalPhysicalMemory());
 }
 #endif
